+++ NEWS.md
@@ -1,4 +1,13 @@
 # battleRealms 0.0.0.9000
 
-* Added a `NEWS.md` file to track changes to the package.
+## Data
+
+* Battle Realms data avialable are `br_abilities`, `br_battle_gears`, `br_buildings`, `br_team_colors`, `br_techniques`, `br_unit_stats`, `br_units`, and `br_weapons`. Documentation to follow.
+
+## Vignettes
+
 * Working on first vignette for a basic unit strenght, weaknesses, and statistics for quick reference.
+
+## Miscellaneuos
+
+* Added a `NEWS.md` file to track changes to the package.